chore: adjust SKCanvasElement_Simple and other adjustments
