+++ src/AddIns/Uno.WinUI.Graphics2DSK/Uno.WinUI.Graphics2DSK.csproj
@@ -1,23 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-	<Choose>
-		<When Condition="$(UnoTargetFrameworkOverride.StartsWith('net9.0'))">
-			<PropertyGroup>
-				<TargetFrameworks>$(NetCurrent);$(NetWinAppSDKPreviousAndCurrent)</TargetFrameworks>
-			</PropertyGroup>
-		</When>
-		<When Condition="$(UnoTargetFrameworkOverride.StartsWith('net8.0'))">
-			<PropertyGroup>
-				<TargetFrameworks>$(NetPrevious);$(NetWinAppSDKPreviousAndCurrent)</TargetFrameworks>
-			</PropertyGroup>
-		</When>
-		<When Condition="$(UnoTargetFrameworkOverride) != ''">
-			<PropertyGroup>
-				<TargetFrameworks>$(NetSkiaPreviousAndCurrent);$(NetWinAppSDKPreviousAndCurrent)</TargetFrameworks>
-			</PropertyGroup>
-		</When>
-	</Choose>
+	<PropertyGroup>
+		<TargetFrameworks>$(NetSkiaPreviousAndCurrent);$(NetMobilePreviousAndCurrent);$(NetWinAppSDKPreviousAndCurrent)</TargetFrameworks>
+	</PropertyGroup>
+	<Import Project="../../targetframework-override.props" />
 
 	<PropertyGroup>
+		<AssemblyName>Uno.WinUI.Graphics2DSK</AssemblyName>
 		<RootNamespace>Uno.WinUI.Graphics2DSK</RootNamespace>
 		<GeneratePackageOnBuild Condition="'$(UNO_UWP_BUILD)'!='true' And '$(Configuration)'=='Release'">true</GeneratePackageOnBuild>
 
@@ -29,24 +17,39 @@
 		<PackageReference Include="SkiaSharp" />
 	</ItemGroup>
 
-	<ItemGroup Condition="!$(TargetFramework.Contains('windows10'))">
-		<!-- This project will be referenced by other projects that reference Uno.UI.Skia and Uno.UI.Composition.Skia
-	 	 so it's important to add the PrivateAssets part below to keep the references here from competing
-	 	 against those references -->
-		<ProjectReference Include="..\..\Uno.UI\Uno.UI.Reference.csproj" PrivateAssets="all" />
-		<ProjectReference Include="..\..\Uno.UI.Composition\Uno.UI.Composition.Reference.csproj" PrivateAssets="all" />
-	</ItemGroup>
-	<PropertyGroup Condition="$(TargetFramework.Contains('windows10'))">
-		<RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
-		<DefineConstants>$(DefineConstants);WINAPPSDK</DefineConstants>
-		<UseRidGraph>true</UseRidGraph>
-		<WindowsPackageType>None</WindowsPackageType>
-		<EnableMsixTooling>true</EnableMsixTooling>
-	</PropertyGroup>
-	<ItemGroup Condition="$(TargetFramework.Contains('windows10'))">
-		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.5.240227000" />
-		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.756" PrivateAssets="all" />
-	</ItemGroup>
+	<Choose>
+		<When Condition="$(TargetFramework.Contains('windows10'))">
+			<PropertyGroup>
+				<RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
+				<UseRidGraph>true</UseRidGraph>
+				<WindowsPackageType>None</WindowsPackageType>
+				<EnableMsixTooling>true</EnableMsixTooling>
+			</PropertyGroup>
+			<ItemGroup>
+				<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.5.240227000" />
+				<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.756" PrivateAssets="all" />
+			</ItemGroup>
+		</When>
+		<When Condition="$(TargetFramework.Contains('android'))">
+			<ItemGroup>
+				<ProjectReference Include="..\..\Uno.UI\Uno.UI.netcoremobile.csproj" PrivateAssets="all" />
+				<ProjectReference Include="..\..\Uno.UI.Composition\Uno.UI.Composition.netcoremobile.csproj" PrivateAssets="all" />
+			</ItemGroup>
+		</When>
+		<Otherwise>
+			<PropertyGroup>
+				<DefineConstants>$(DefineConstants);REFERENCE</DefineConstants>
+			</PropertyGroup>
+			<ItemGroup>
+				<!-- This project will be referenced by other projects that reference Uno.UI.<platform> and Uno.UI.Composition.<platform>
+				 so it's important to add the PrivateAssets part below to keep the references here from competing
+				 against those references -->
+				<ProjectReference Include="..\..\Uno.UI\Uno.UI.Reference.csproj" PrivateAssets="all" />
+				<ProjectReference Include="..\..\Uno.UI.Composition\Uno.UI.Composition.Reference.csproj" PrivateAssets="all" />
+			</ItemGroup>
+		</Otherwise>
+	</Choose>
+	
 
 	<ItemGroup>
 		<None Include="..\..\Common\uno.png" Pack="true" PackagePath="\" />
