+++ src/SamplesApp/SamplesApp.Wasm/SamplesApp.Wasm.csproj
@@ -114,6 +114,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\..\AddIns\Uno.UI.Lottie\Uno.UI.Lottie.Wasm.csproj" />
+		<ProjectReference Include="..\..\AddIns\Uno.WinUI.Graphics2DSK\Uno.WinUI.Graphics2DSK.csproj" />
 		<ProjectReference Include="..\..\AddIns\Uno.UI.MediaPlayer.WebAssembly\Uno.UI.MediaPlayer.WebAssembly.csproj" />
 		<ProjectReference Include="..\..\AddIns\Uno.UI.MSAL\Uno.UI.MSAL.Wasm.csproj" />
 		<ProjectReference Include="..\..\Uno.Foundation.Runtime.WebAssembly\Uno.Foundation.Runtime.WebAssembly.csproj" />
