+++ src/Uno.UI/Graphics/SKCanvasVisualBase.cs
